Add new cambot module to store code base in. Split dynamic playlister into its own file
