+++ cambot/lastfm_song_count_given_daterange.py
@@ -6,7 +6,7 @@ import pylast
 import sys
 from twython import Twython
 import argparse
-import config
+import cambot.config as config
 
 # Twitter API
 twitter_api_key = config.twitter_api_key
