+++ cambot/top_tracks_playlister.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python
 import datetime
-import operator
 
 import spotipy
 from twython import Twython
@@ -154,94 +153,11 @@ def check_number(value):
         raise argparse.ArgumentTypeError("%s is an invalid int value. The limit is between 1 and 50" % value)
     return ivalue
 
-
-def dynamic_playlist_updater(playlist_id):
-    # Start with getting the top songs for each artist stored in the database
-    mydb = myclient["artist"]
-    mycol = mydb["collection"]
-
-    top_tracks = [user.get_top_tracks(period="7day"),
-                  user.get_top_tracks(period="1month"),
-                  user.get_top_tracks(period="3month"),
-                  user.get_top_tracks(period="6month"),
-                  user.get_top_tracks(period="12month"),
-                  user.get_top_tracks(period="overall")]
-
-    # TODO: user.getRecentTracks with time_ranges. Use same time range song weights. e.g. 1 week all 100 for each hit. 1 month (minus that week thats already been counted)
-
-    # Loop over all the artists in the database
-    collection_of_artists = mycol.find()
-    for artist in collection_of_artists:
-        # Find this artist in each top songs list. Shorter time frame listens count as a higher point score than one from
-        # a time period that is longer. E.g. 1 listen from the 7day period = 100 points. 1 listen from 1 month period = 60 and so on
-        artist_name = artist["name"]
-        reference_count = 1
-        list_of_songs_and_their_weight = defaultdict(int)
-        for period_top_tracks_list in top_tracks:
-            for track in period_top_tracks_list:
-                track_and_artist = str(track[0])
-                scrobbles = int(track[1])
-                # Add the song to the dict if not exist, and either way update that songs weighting
-                list_of_songs_and_their_weight[track_and_artist] += (calculate_song_weight(reference_count) * scrobbles)
-            reference_count += 1
-        reference_count = 1
-        sorted_tracks = OrderedDict(sorted(list_of_songs_and_their_weight.items(), key=operator.itemgetter(1), reverse=True))
-
-        if token:
-            sp = spotipy.Spotify(auth=token)
-
-
-            # Get a list of tracks from sorted list
-            spotify_searched_tracks = []
-            limit = 0
-            for key, value in sorted_tracks.items():
-                if limit < 50:
-                    search = t.search_spotify(sp, key, "track")
-                    if search is not None:
-                        spotify_searched_tracks.append(search)
-                        limit += 1
-                else:
-                    break
-
-            tracks_to_remove = []
-            tracks_to_add = []
-
-            playlist = sp.user_playlist_tracks(user=spotify_username, playlist_id=playlist_id)
-            # order = 0
-            # for item in playlist['items']:
-            #     link = item["external_urls"].get("url")
-            #     if link not in spotify_searched_tracks:
-            #         # Remove from playlist at end
-            #         tracks_to_remove.append(link)
-            sp.user_playlist_replace_tracks(spotify_username,playlist_id,spotify_searched_tracks)
-
-
-            #sp.user_playlist_remove_all_occurrences_of_tracks(user=spotify_username, playlist_id=playlist_id, tracks=tracks_to_remove)
-
-
-
-def calculate_song_weight(reference_count):
-    if reference_count == 1:
-        points_per_hit = 200
-    elif reference_count == 2:
-        points_per_hit = 90
-    elif reference_count == 3:
-        points_per_hit = 35
-    elif reference_count == 4:
-        points_per_hit = 15
-    elif reference_count == 5:
-        points_per_hit = 7
-    elif reference_count == 6:
-        points_per_hit = 5
-    return points_per_hit
-
-
 parser = argparse.ArgumentParser()
 parser.add_argument("-t", "--timeframe",
                     help="Creates a playlist of the top songs from the past month, 6 months or years.",
                     choices=['short_term', 'medium_term', 'long_term'])
 parser.add_argument("--limit", help="Specify how many tracks you want in the playlist", default=50, type=check_number)
-parser.add_argument("-d", "--dynamic", help="Runs the dynamic playlist updater.", default="09mHNJgzOgdUZksMrohoPH")
 parser.add_argument("--tweet",
                     help="If this has been entered, then the script will NOT tweet to the account provided.",
                     action="store_true")
@@ -254,7 +170,6 @@ time_range = args.timeframe
 limit = int(args.limit)
 add = args.at
 tweet = args.tweet
-dynamic = args.dynamic
 
 if add == 'add':
     print("adding artist names to the database.", flush=True)
@@ -264,8 +179,4 @@ if tweet:
     print("TEST MODE: not tweeting to the account provided")
     TWEET = False
 
-if dynamic:
-    print("Dynamic playlist mode selected", flush=True)
-    dynamic_playlist_updater(dynamic)
-else:
-    create_playlist(time_range, limit)
+create_playlist(time_range, limit)
